DoE.base-package {DoE.base}	R Documentation

Full factorials, orthogonal arrays and base utilities for DoE packages

Description

This package creates full factorial designs and designs from orthogonal arrays. In addition, it provides some basic utilities like an exporting function for the DoE packages FrF2, DoE.wrapper and RcmdrPlugin.DoE, and some diagnostics for general orthogonal arrays (generalized word length calculations).

Details

The package is still in under development phase. Details about combining designs are particularly likely to be changed in the future (param.design, cross.design). Please contact me, if you have suggestions.

This package designs full factorial experiments (function fac.design) and experiments based on orthogonal arrays (oa.design). Some aspects of functions fac.design and oa.design have been modeled after the functions of the same name given in Chambers and Hastie (1993) (e.g. for the option factor.names or for outputting a data frame with attributes). However, S compatibility has not been considered in devising this package.

The orthogonal arrays underlying function oa.design are mainly taken from Kuhfeld (2009). While the arrays generally guarantee estimability of main effects in case there are no (or negligible) active interactions, some of them can also be used for designs for which some interactions are to be estimated, if only few of the design columns are used for experimentation. Optimization for such purposes and check of fitness for such purposes is supported, cf. generalized.word.length.

The package provides class design for use also by packages FrF2, DoE.wrapper and RcmdrPlugin.DoE. Furthermore, it provides utilities for printing, plotting, summarizing, exporting and combining experimental designs. Package FrF2 relies on function fac.design for full factorials in 2-level factors.

Acknowledgments

Thanks are due to Peter Theodor Wilrich for various useful suggestions in the early phase of this project!

Author(s)

Ulrike Groemping

Maintainer: Ulrike Groemping <groemping@bht-berlin.de>

References

Chambers, J.M. and Hastie, T.J. (1993). Statistical Models in S, Chapman and Hall, London.

Groemping, U. (2018). R Package DoE.base for Factorial Designs. Journal of Statistical Software 85(5), 1–41. https://www.jstatsoft.org/issue/view/v085.

Hedayat, A.S., Sloane, N.J.A. and Stufken, J. (1999) Orthogonal Arrays: Theory and Applications, Springer, New York.

Kuhfeld, W. (2009). Orthogonal arrays. Website courtesy of SAS Institute http://support.sas.com/techsup/technote/ts723.html.

See Also

Functions fac.design, oa.design for generating designs, and various functions (generalized.word.length) for optimizing and checking a designs properties,
class design which is utilized also by packages FrF2 and DoE.wrapper.
Furthermore, there are various utility functions like export.design or add.response and functions cross.design or param.design for combining designs.
Finally, several methods for class design objects are provided, especially also functions formula.design and lm.design for automatic generation of linear models (but beware: these are convenience functions that provide a quick first look but NOT necessarily the best statistical approach to analysis!).

[Package DoE.base version 1.1-4 Index]