GRind {DoE.base}	R Documentation

Functions for calculating generalized resolution, average R-squared values and squared canonical correlations, and for checking design regularity

Description

Function GR calculates generalized resolution, function GRind calculates more detailed generalized resolution values, squared canonical correlations and average R-squared values, the print method for class GRind appropriately prints the detailed GRind values. Function SCFTs calculates squared canonical correlations for factorial designs. SCFTs includes more projections than GRind (all full resolution projections or even all projections) and decides on regularity of the design, based on a conjecture.

Usage

GR(ID, digits=2)
GRind(design, digits=3, arft=TRUE, scft=TRUE, cancors=FALSE, with.blocks=FALSE)
SCFTs(design, digits = 3, all = TRUE, resk.only = TRUE, kmin = NULL, kmax = ncol(design),
   regcheck = FALSE, arft = TRUE, cancors = FALSE, with.blocks = FALSE)
## S3 method for class 'GRind'
print(x, quote=FALSE,  ...)

Arguments

ID
an orthogonal array, either a matrix or a data frame; need not be of class oa; can also be a character string containing the name of an array listed in data frame oacat

digits
number of decimal points to which to round the result

design
a factorial design. This can either be a matrix or a data frame in which all columns are experimental factors, or a special data frame of class design, which may also include response data.
In any case, the design should be a factorial design; the functions are not useful for quantitative designs (like e.g. latin hypercube samples).

arft
logical indicating whether or not the average $R^2$ frequency table (ARFT, see Gr?mping 2013) is to be returned

scft
logical indicating whether the squared canonical correlation frequency table (SCFT, see Gr?mping 2013) is to be returned

cancors
logical indicating whether individual canonical correlations are to be returned (see Groemping 2013). These will not be needed for normal use of the package.

with.blocks
a logical, indicating whether or not an existing block factor is to be included into word counting. This option is ignored if design is not of class design.
Per default, an existing block factor is ignored.
For designs without a block factor, the option does not have an effect.
If the design is blocked, and with.blocks is TRUE, the block factor is treated like any other factor in terms of word counting.

all
logical; decides whether or not to consider projections of more than R~factors, where R denotes the design resolution

resk.only
logical; if all is TRUE, should only full resolution projections be considered? Choosing FALSE may cause very long run times.

kmin
integer; purpose is to continue an earlier run with additional larger projections

kmax
integer; limit on projection sizes to consider

regcheck
logical; is the purpose a regularity check? If TRUE, the function stops after the first projection size that included squared canonical correlation different from 0 or 1.

x
a list of class GRind, as created by function GRind

quote
a logical indicating whether character values are quoted

...
further arguments to function print

Details

Functions GR, GRind, and SCFTs work for factors only and are not intended for quantitative variables. Nevertheless it is possible to apply them to class design plans with quantitative variables in them in some situations.

Function GR calculates the generalized resolution according to Deng and Tang (1999) for 2-level designs or a generalization thereof according to Groemping (2011) and Groemping and Xu (2014) for general orthogonal arrays. It returns a value between 3 and 5, where the numeric value 5 stands for “at least 5”. Roughly, generalized resolution measures the closeness of a design to the next higher resolution (worst-case based, e.g. one completely aliased triple of factors implies resolution 3).

Function GRind (newer than GR, and recommended) calculates the generalized resolution, together with factor wise generalized resolution values, squared canonical correlations and average R-squared values, as mentioned in Groemping and Xu (2014) and further developed in Groemping (2013, 2017). The print method for class Grind objects prints the individual factor components of GRind.i such that they do not mislead: Because of the shortest word approach for GR, SCFT and ARFT, a GRind.i component can be at most one larger than the resolution. For example, if GR is 3.5 so that the resolution is 3, the largest possible numeric value of a GRind.i component is 4, but it means ">=4".

Function SCFTs does more extensive SCFT and ARFT calculations than function GRind: in particular, the function allows to do such calculations for more projection sizes, either restricting attention to full resolution projections or going for ALL projections with non-zero word lengths. These capabilities have been introduced in relation to regularity checking based on SCFTs (see Groemping and Bailey 2016): Defining a factorial design as regular if all main effects are orthogonal in some sense to effects including other factors of any order, it is conjectured that a regularity check on full resolution projections only will suffice for identifying non-regularity (work in progress). However, this is a conjecture only; as long as it is not proven, a definite check for this type of regularity requires checking ALL projections, i.e. setting resk.only to FALSE. With this setting, the function may run for a very long time (depends in particular on the number of factors)!

Value

Function GR returns a list with elements GR (the generalized resolution of the array, a not necessarily integer number between 3 and 5) and RPFT (the relative projection frequency table). GR values smaller than 5 are exact, while the number five stands for “at least 5”. The resolution itself is the integer portion of GR. The RPFT element is the relative projection frequency table for 4-factor projections for GR=5. For unconfounded three- and four-column designs, GR takes the value Inf (used to be 5 for package versions up to 0.23-4).

Function GRind works on designs with resolution at least 3 and returns a list with elements
GRs (the two versions of generalized resolution described in Groemping and Xu 2014),
the matrix GR.i with rows GRtot.i and GRind.i for the factor wise generalized resolutions (also in Groemping and Xu 2014),
and optionally
the ARFT (Groemping 2013, 2017),
the SCFT (Groemping 2013, 2017),
and/or the canonical correlations.
The latter are held in an
nfac x choose(nfac-1, R-1) x max(nlev)-1 array
and are supplemented with 0es, if there are fewer of them than the respective dfi.

The factor wise generalized resolutions are in the closed interval between resolution and resolution + 1. In the latter case, their meaning is "at least resolution + 1". (The print method ensures that they are printed accordingly, but the list elements themselves are just the numbers.)

Function SCFTs returns a list of lists with a component for each projection size considered. Each such component contains the following entries:

SCFT
Squared canonical correlation table for the projection size

ARFT
Average R^2 frequency table for the projection size (if requested)

cancors
canonical correlations (if requested)

Warning

The functions have been checked on the types of designs for which they are intended (especially orthogonal arrays produced with oa.design) and on 2-level fractional factorial designs produced with package FrF2. They may produce meaningless results for some other types of designs.

Author(s)

Ulrike Groemping

References

Groemping, U. (2011). Relative projection frequency tables for orthogonal arrays. Report 1/2011, Reports in Mathematics, Physics and Chemistry http://www1.beuth-hochschule.de/FB_II/reports/welcome.htm, Department II, Beuth University of Applied Sciences, Berlin.

Groemping, U. (2013). Frequency tables for the coding invariant ranking of orthogonal arrays. Report 2/2013, Reports in Mathematics, Physics and Chemistry http://www1.beuth-hochschule.de/FB_II/reports/welcome.htm, Department II, Beuth University of Applied Sciences, Berlin.

Groemping, U. (2017). Frequency tables for the coding invariant quality assessment of factorial designs. IISE Transactions 49, 505-517. https://doi.org/10.1080/0740817X.2016.1241458.

Groemping, U. and Bailey, R.A. (2016). Regular fractions of factorial arrays. In: mODa 11 – Advances in Model-Oriented Design and Analysis. New York: Springer.

Groemping, U. and Xu, H. (2014). Generalized resolution for orthogonal arrays. The Annals of Statistics 42, 918–939. https://projecteuclid.org/euclid.aos/1400592647

See Also

See also GWLP and generalized.word.length

Examples

   oa24.bad <- oa.design(L24.2.13.3.1.4.1, columns=c(1,2,14,15))
   oa24.good <- oa.design(L24.2.13.3.1.4.1, columns=c(3,10,14,15))
   ## generalized resolution differs (resolution is III in both cases)
   GR(oa24.bad)
   GR(oa24.good)

   ## and analogously also GRind and ARFT and SCFT
   GRind(oa24.bad)
   GRind(oa24.good)

   ## SCFTs
   ## Not run: plan <- L24.2.12.12.1[,c(1:5,13)]
   GRind(plan)  ## looks regular (0/1 SCFT only)
   SCFTs(plan)
   SCFTs(plan, resk.only=FALSE)
   
## End(Not run)

[Package DoE.base version 1.1-4 Index]