SN {DoE.base}	R Documentation

Function for the signal-to-noise ratio 10 * log10(mean^2/var)

Description

Function for the signal-to-noise ratio 10 * log10(mean^2/var)

Usage

SN(x)

Arguments

x
a data vector to take the S/N ratio over

Details

Taguchi proposes three different versions of S/N-ratio. In line with Box, Hunter and Hunter (2005), only the one for target-optimization is given here, as it is invariant against linear transformation.

Value

a number (10 * log10(mean^2/var))

Note

This package is currently under intensive development. Substantial changes are to be expected in the near future.

Author(s)

Ulrike Groemping

References

Box G. E. P, Hunter, W. C. and Hunter, J. S. (2005) Statistics for Experimenters, 2nd edition. New York: Wiley.

See Also

See also aggregate.design; function SN has been developed for use with aggregating parameter designs

Examples

x <- rexp(10)
SN(x)
10 * log10(mean(x)^2/var(x))
20 * log10(mean(x)/sd(x))

[Package DoE.base version 1.1-1 Index]