block.catlg3 {DoE.base}	R Documentation

Catalogues for blocking full factorial 2-level and 3-level designs, and lists of generating columns for regular 2- and 3-level designs.

Description

The block data frames hold Yates matrix column numbers for blocking full factorials with 2-level (up to 256 runs) and 3-level factors (up to 243 runs). The Yates lists translate these column numbers into effects.

Usage

block.catlg
block.catlg3
Yates
Yates3

Details

The constants documented here are used for blocking full factorial designs with function fac.design; Yates and block.catlg are internal here, as they have long been part of package FrF2-package.

The block data frames hold Yates matrix column numbers for blocking full factorials with 2-level (up to 256 runs) and 3-level factors (up to 243 runs). The Yates lists translate these column numbers into effects (see below).

Data frame block.catlg comes from Sun, Wu and Chen (1997). Data frame block.catlg3 comes from Cheng and Wu (2002, up to 81 runs) and has been derived from Hinkelmann and Kempthorne (2005, Table 10.6) for 243 runs. The blocking schemes from the papers are optimal; this has not been proven for the blocking scheme for 243 runs.

Yates is a user-visible constant that is useful in design construction:

Yates is a list of design column generators in Yates order (for 4096 runs), e.g. Yates[1:8] is identical to

list(1,2,c(1,2),3,c(1,3),c(2,3),c(1,2,3)).

Yates3 is a constant for 3-level designs, for which there are coefficients rather than generating factor numbers in the list.

Author(s)

Ulrike Groemping

References

Cheng, S.W. and Wu, C.F.J. (2002). Choice of Optimal Blocking Schemes in Two-Level and Three-Level Designs. Technometrics 44, 269-277.

Hinkelmann, K. and Kempthorne, O. (2005). Design and analysis of experiments, Vol.2. Wiley, New York.

Sun, D.X., Wu, C.F.J. and Chen, Y.Y. (1997). Optimal blocking schemes for 2\^p and 2\^(n-p) designs. Technometrics 39, 298-307.

[Package DoE.base version 1.1-1 Index]