contr.FrF2 {DoE.base}	R Documentation

Contrasts for orthogonal Fractional Factorial 2-level designs

Description

Contrasts for orthogonal Fractional Factorial 2-level designs

Usage

   contr.FrF2(n, contrasts=TRUE)

Arguments

n
power of 2; number of levels of the factor for which contrasts are to be generated

contrasts
must always be TRUE; option needed for function model.matrix to work properly

Details

This function mainly supports -1/+1 contrasts for 2-level factors. It does also work if the number of levels is a power of 2. For more than four levels, the levels of the factor must be in an appropriate order in order to guarantee that the columns of the model matrix for an FrF2-derived structure are orthogonal.

Value

The function returns orthogonal contrasts for factors with number of levels a power of 2. All contrast columns consist of -1 and +1 entries (half of each). If factors in orthogonal arrays with 2-level factors are assigned these contrasts, the columns of the model matrix for the main effects model are orthogonal to each other and to the column for the intercept.

Note

This package is currently under intensive development. Substantial changes are to be expected in the near future.

Author(s)

Ulrike Groemping

See Also

See Also contrasts, FrF2, fac.design, oa.design, pb

Examples

## assign contr.FrF2 contrasts to a factor
status <- as.factor(rep(c("current","new"),4))
contrasts(status) <- contr.FrF2(2)
contrasts(status)

[Package DoE.base version 1.1-1 Index]