genChild {DoE.base}	R Documentation

Internal utility functions to support automatic creation of child arrays from entries of the data frame oacat

Description

The functions are used internally for creating the child arrays listed in data frame oacat from the parent arrays that come with DoE.base (or using full factorials).

Usage

parseArrayLine(array.line)
genChild(array.list)
getArray(nbRuns, descr)
symb2oa(nbRuns, descr)
oa2symb(name)

Arguments

array.line
a row from data frame oacat

array.list
the output from function parseArrayLine

nbRuns
the number of runs of the array to be received

descr
a character string containing the description of the array to be retrieved, of the form
nl1~fl1;nl2~fl2; ..., where nl stands for the number of levels and fl for their respective frequency; the string may (but need not) contain a trailing semicolon

name
name of an array according to the naming conventions in oacat

Details

Function parseArrayLine transforms information from a row of oacat into a list format digestible by function genChild.

Function genChild creates a child array from the appropriate information provided by function parseArrayLine.

Function getArray retrieves a stored orthogonal array based on the list information it receives.

Functions symb2oa and oa2symb can be used for switching between design names from data frame oacat and list type information used by functions internally. Note that the result from oa2symb is not sufficient to get back to the oa representation, but needs the number of runs in addition.

Value

parseArrayLine returns a list with design and lineage description in symbolic form.

genChild and getArray return an array matrix of class oa.

symb2oa and oa2symb return a character string.

Author(s)

Boyko Amarov and Ulrike Groemping

See Also

See also oacat, oa

[Package DoE.base version 1.1-1 Index]