iscube {DoE.base}	R Documentation

Functions to isolate cube points from 2-level fractional factorial design with center and / or star points

Description

These functions identify the positions for cube points or star points and can reduce a central composite design to its cube portion (with center points).

Usage

iscube(design, ...)
isstar(design, ...)
pickcube(design, ...)

Arguments

design
a data frame of class design that contains a 2-level fractional factorial (regular or non-regular) or a central composite design.

...
currently not used

Details

Function iscube provides a logical vector that is TRUE for cube points and FALSE for center points and star points. Its purpose is to enable use of simple functions for “clean” 2-level fractional factorials like MEPlot or DanielPlot.

Function isstar provides a logical vector that is TRUE for the star block (including center points) of a central composite design.

Function pickcube reduces a central composite design (type ccd) to its cube block, including center points. This function is needed, if a CCD has been created in one go, but analyses are already required after conducting the cube portion of the design (and these perhaps even prevent the star portion from being run at all).

Value

iscube and isstar each return a logical vector (cf. Details section).

pickcube returns a data frame of class design with type FrF2.center or FrF2.

Warning

For version 0.22-8 of package DoE.base, function iscube returned a wrong result without warning, when applied to an old version CCD design (before DoE.wrapper, version 0.8-6 of Nov 15 2011).
Since version 0.23 of package DoE.base, the function works on old designs, except for blocked or replicated versions; for these, an error is thrown.)

Note

The functions have not been tested for central composite designs for which the cube portion itself is blocked.

Author(s)

Ulrike Groemping

References

Montgomery, D.C. (2001). Design and Analysis of Experiments (5th ed.). Wiley, New York.

See Also

See also as pb, FrF2, ccd.design

Examples

  ## purely technical example, not run because FrF2 not loaded
  ## Not run: 
  plan <- FrF2(16,5, factor.names=c("one","two","three","four","five"), ncenter=4)
  iscube(plan)
  plan2 <- ccd.augment(plan)
  iscube(plan2)
  isstar(plan2)
  pickcube(plan2)
  
## End(Not run)
   

[Package DoE.base version 1.1-4 Index]