lowerbound_AR {DoE.base}	R Documentation

Function to Calculate a Lower Bound for A_R and Internal Auxiliary Functions

Description

The functions serve the calculation of lower bounds for the worst case confounding. lowerbound_AR is intended for direct use, lowerbounds and lowerbound_chi2 are internal functions.

Usage

lowerbound_AR(nruns, nlevels, R, crit = "total")
lowerbounds(nruns, nlevels, R)
lowerbound_chi2(nruns, nlevels)

Arguments

nruns
positive integer, the number of runs

nlevels
vector of positive integers, the numbers of levels for the factors

R
positive integer, the resolution of the design; if it is uncertain whether resolution R is feasible, this should be checked by function oa_feasible before applying any of the lower bound functions.

crit
"total" or "worst"; if "total", a bound for the overall A_R (sum of the results from lowerbounds) is calculated; otherwise, a bound for the largest individual contribution from an R factor set is calculated

Details

Note: if the specified resolution R is not feasible (necessary conditions can be checked with function oa_feasible), any bound(s) returned will be meaningless.

Function lowerbounds provides (integral) bounds on n^2 A_R (with n=nruns) according to Groemping and Xu (2014) Theorem 5 for all R factor sets. If the number of runs permits a design with resolution larger than R, the value(s) will be 0. For resolution at least III, the result of function lowerbound_AR is the sum (crit="total") or maximum (crit="worst") of these individual bounds, divided by the square of the number of runs.

For resolution II and crit="total", function lowerbound_chi2 implements the lower bound B on chi^2 which was provided in Lemma 2 of Liu and Lin (2009). For supersaturated resolution II designs, this bound is is usually sharper than the one obtained on the basis of Grömping and Xu (2014). Due to the relation between A_2 and chi^2 that is stated in Groemping (2017) (summands of A_2 are an nth of a chi^2, with n=nruns), this bound can be easily transformed into a bound for A_2; this relation is also used to slightly sharpen the bound B itself: n^2 A_2 must be integral, which implies that B can be replaced by ceiling(nruns*B)/nruns, which is applied in function lowerbound_chi2. Function lowerbound_AR increases the lower bound on A_2 accordingly, if lowerbound_chi2 provides a sharper bound than the sum of the elements returned by functioni lowerbounds.

Value

lowerbound_AR returns a lower bound for the number of words of length R (either total or worst case),
lowerbounds returns a vector of lower bounds for individual R factor sets on a different scale (division by nruns^2 needed for transforming this into the contributions to words of length R),
and function lowerbound_chi2 returns a lower bound on the chi^2 value which can be used as a quality criterion for supersaturated designs.

Author(s)

Ulrike Groemping

References

Groemping, U. and Xu, H. (2014). Generalized resolution for orthogonal arrays. The Annals of Statistics 42, 918-939.

Groemping, U. (2017). Frequency tables for the coding-invariant quality assessment of factorial designs. IISE Transactions 49, 505-517.

Liu, M.Q. and Lin, D.K.J. (2009). Construction of Optimal Mixed-Level Supersaturated Designs. Statistica Sinica 19, 197-211.

See Also

See also oa_feasible.

Examples

lowerbound_AR(24, c(2,3,4,6),2)

[Package DoE.base version 1.1-4 Index]