print.oa {DoE.base}	R Documentation

Function to Print oa Objects with a Lot of Added Info

Description

The function suppresses printing of voluminous info attached as attributes to oa objects.

Usage

## S3 method for class 'oa'
print(x, ...)

Arguments

x
the oa object to be printed

...
further arguments for default print function

Details

The function currently suppresses the printing of the attributes MIPinfo, history and matrices, if present; a small excerpt of the MIPinfo attribute is printed. Additionally, the names of unusual attributes are printed. They can also be printed separately by running names(attributes(x)); to access an attribute, run attr(x, "MIPinfo"), for example.

Value

The function is used for its side effects and does not return anything.

Author(s)

Ulrike Groemping

See Also

See also print.default and str

[Package DoE.base version 1.1-1 Index]