show.oas {DoE.base}	R Documentation

Function to display list of available orthogonal arrays

Description

This function allows to inspect the list of available orthogonal arrays, optionally specifying selection criteria

Usage

show.oas(name = "all", nruns = "all", nlevels = "all", factors = "all", 
    regular = "all", GRgt3 = c("all", "tot", "ind"), Rgt3 = FALSE, show = 10, 
    parents.only = FALSE, showGRs = FALSE, showmetrics = FALSE, digits = 3)

Arguments

name
character string or vector of character strings giving name(s) of (an) orthogonal array(s); results in an error if name does not contain any valid name; warns if name contains any invalid name

nruns
the requested number of runs or a 2-element vector with a minimum and maximum for the number of runs

nlevels
a vector of requested numbers of levels for a set of factors in question, must contain integers > 1 only;
nlevels cannot be specified together with factors

factors
a list with the two elements nlevels and number, which are both integer vectors of equal length;
nlevels contains the number of levels and number the number of factors for the corresponding number of levels

regular
either unrestricted (the default “all”), a logical which requests (TRUE) or rejects (FALSE) regular arrays, or the character string “strict” to request strictly regular arrays, for which all confounded factors are completely confounded with a 2-factor interaction of two other factors (the latter are fixed level arrays or crossed arrays)

GRgt3
either unrestricted (the default “all”), or a character string which requests GR ("tot") or GRind ("ind") to be larger than 3

Rgt3
logical requesting inclusion of standard resolution 3 arrays as listed in oacat per default, and restricting the output to arrays of resolution at least IV (as listed in oacat3), if changed to TRUE

show
an integer number specifying how many arrays are to be listed (upper bound), or the character string "all" for showing all arrays, no matter how many. The default is to show 10 arrays. show = 0 switches off the display of the result and only returns a value. Since August 2018, the number refers to stronger and weaker arrays, separately.

parents.only
logical specifying whether to show only parent arrays or child arrays as well; the default is FALSE for inclusion of child arrays

showGRs
logical specifying whether to show the generalized resolution quality metrics with the resulting arrays; the default is FALSE. If set to TRUE, three metrics are displayed (see Details section).

showmetrics
logical specifying whether to show all array quality metrics with the resulting arrays; the default is FALSE. If set to TRUE, several metrics are displayed (see Details section).

digits
integer number of significant digits to show for GR and A metrics; irrelevant, if showmetrics is FALSE

Details

The function shows the arrays that are listed in the data frames oacat or oacat3.

For child arrays that have to be generated with a lineage rule (can be automatically done with function oa.design), the lineage is displayed together with the array name. The option parent.only = TRUE suppresses printing and output of child arrays. The structure of the lineage entry is documented under oacat.

If display of metrics is requested with showmetrics=TRUE, the printed output shows the metrics GR*, GRind*, regular (logical, whether regular or not), SCones* (number of squared canonical correlations that are 1), and the numbers of words of lengths 3 to 8 (A3 to A8). showGRs=TRUE requests the metrics marked with asterisks only (without SCones in case GRgt3="ind"). More information on all these metrics can be found here

Value

A data frame with the three columns name, nruns and lineage, containing the array name, the number of runs and - if applicable - the lineage for generating the array from other arrays. The lineage entry is empty for parent arrays that are either directly available in the package and can be accessed by giving their name (e.g. L18.3.6.6.1) or are full factorials (e.g. L28.4.1.7.1). If further information has been requested (e.g. with showmetrics=TRUE), the data frame contains additional columns.

If no array has been found, the returned value is NULL.

Note

Thanks to Peter Theodor Wilrich for proposing such a function.

Author(s)

Ulrike Groemping

References

Kuhfeld, W. (2009). Orthogonal arrays. Website courtesy of SAS Institute http://support.sas.com/techsup/technote/ts723.html.

Mee, R. (2009). A Comprehensive Guide to Factorial Two-Level Experimentation. New York: Springer.

See Also

oa.design for using the arrays from oacat in design creation
oacat for the data frames underlying the function

Examples

   ## the first 10 orthogonal arrays with 24 to 28 runs
   show.oas(nruns = c(24,28))
   ## the first 10 orthogonal arrays with 24 to 28 runs
   ## excluding child arrays
   show.oas(nruns = c(24,28), parents.only=TRUE)
   ## the orthogonal arrays with 4 2-level factors, one 4-level factor and one 5-level factor
   show.oas(factors = list(nlevels=c(2,4,5),number=c(4,1,1)))
   ## show them all with quality metrics
   show.oas(factors = list(nlevels=c(2,4,5),number=c(4,1,1)), show=Inf, showmetrics=TRUE)
   ## pick only those with no complete confounding of any degrees of freedom
   show.oas(factors = list(nlevels=c(2,4,5),number=c(4,1,1)), GRgt3="ind", showmetrics=TRUE)
   ## the orthogonal arrays with 4 2-level factors, one 7-level factor and one 5-level factor
   show.oas(factors = list(nlevels=c(2,7,5),number=c(4,1,1)))
   ## the latter orthogonal arrays with the nlevels notation 
   ## (that can also be used in a call to oa.design subsequently)
   show.oas(nlevels = c(2,7,2,2,5,2))
   ## calling arrays by name 
   show.oas(name=c("L12.2.11", "L18.2.1.3.7"))

[Package DoE.base version 1.1-4 Index]